append

+++ 0x15-file_io/2-append_text_to_file.c
@@ -9,7 +9,7 @@
 int append_text_to_file(const char *filename, char *text_content)
 {
 	int fd;
-	ssize_t w;
+	int w;
 	int len = 0;
 
 	if (filename == NULL)
